refactor(aiproxy): extract findTokenByName into lib/v2alpha/tokens
Pulls the token lookup helper out of the [name]/route.ts handler so it can
be shared by upcoming enable/disable endpoints. No behavior change.
